Fix build on 64-bit

+++ 32blit/graphics/blend.cpp
@@ -53,7 +53,7 @@ namespace blit {
 
     // if destination is not double-word aligned copy at most three bytes until it is
     uint8_t* de = d + c * 3;
-    while (uint32_t(d) & 0b11) {
+    while (uintptr_t(d) & 0b11) {
       *d = blend((s32 & 0xff), *d, a); d++;
       // rotate the aligned rgbr/gbrg/brgb quad
       s32 >>= 8; s32 |= uint8_t(s32 & 0xff) << 24;
@@ -109,7 +109,7 @@ namespace blit {
 
     // if destination is not double-word aligned copy at most three bytes until it is
     uint8_t* de = d + c * 3;
-    while (uint32_t(d) & 0b11) {
+    while (uintptr_t(d) & 0b11) {
       *d = s32 & 0xff000000; d++;
       // rotate the aligned rgbr/gbrg/brgb quad
       s32 >>= 8; s32 |= uint8_t(s32 & 0xff) << 24;
